Move source loading to seperate file
